requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pypdf
python-dotenv
